#65 - Refactor: 기본 프로필 이미지 컴포넌트화
