+++ src/components/ProfileForm.jsx
@@ -1,11 +1,10 @@
 import "./ProfileForm.css";
-import BasicProfileImg from "../assets/basic-profile-img.svg";
-
+import BasicProfileImg from "./BasicProfileImg";
 export default function ProfileForm() {
     return (
         <form className="profileForm">
             <div className="profileImgSetting">
-                <img src={BasicProfileImg} alt="프로필 이미지" className="profileImg" />
+                <BasicProfileImg size="lg" />
                 <label for="uploadFile" />
                 <input type="file" id="uploadFile" />
             </div>
@@ -23,8 +22,12 @@ export default function ProfileForm() {
                     type="text"
                     placeholder="영문, 숫자, 특수문자(.),(_)만 사용 가능합니다."
                 />
-                <strong className="cautionText">*영문, 숫자, 밑줄 및 마침표만 사용할 수 있습니다.</strong>
-                <strong className="cautionText">*이미 사용 중인 ID입니다.</strong>
+                <strong className="cautionText">
+                    *영문, 숫자, 밑줄 및 마침표만 사용할 수 있습니다.
+                </strong>
+                <strong className="cautionText">
+                    *이미 사용 중인 ID입니다.
+                </strong>
             </div>
             <div className="wrapInput">
                 <label for="">소개</label>
@@ -34,5 +37,5 @@ export default function ProfileForm() {
                 />
             </div>
         </form>
-    )
+    );
 }
